requirements.txt
numpy==1.23.3
llvmlite
numba==0.48.0
#torch>=1.1
tensorboardX
easydict
pyyaml
scikit-image
tqdm
#torchvision
SharedArray==3.0.0
opencv-python
pyquaternion
spconv-cu113
av2
kornia
torch_scatter
shapely==1.8.4
